+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Added --silent and --verbose options for various debug levels.
 
+* Sql::Exporter now support various info & debug messages to better inspect
+  database table alignments and data export tasks.
+
 ## 1.0.0 - 2021-06-09
 
 * First official version used in production.